fix(nvr): build internal/nvr on Windows; guard every package in the cross-compile matrix (#663)
`GOOS=windows go build ./...` failed on main:

    # github.com/aceteam-ai/citadel-cli/internal/nvr
    internal/nvr/storage.go:142:25: undefined: syscall.Stat_t
    internal/nvr/storage.go:234:19: undefined: syscall.Statfs_t

`syscall.Stat`/`Statfs` and their result types do not exist on Windows.

## Why nobody noticed

Releases were unaffected, which is exactly what made it invisible.
build.sh and the CI cross-compile matrix both build `./cmd/citadel`, and
nothing under it imports internal/nvr — only ./cmd/nvrconfig does, and
that is the Linux init container image, built by its own workflow. So
every release and every CI run stayed green while the full-module build
was broken.

The cost is a trap rather than an outage: a contributor checking they
have not broken Windows runs `go build ./...`, hits a failure that is not
theirs, and either burns time on it or learns to ignore that output —
which is how a real Windows break eventually ships.

## Changes

- `internal/nvr/storage_unix.go` (`//go:build !windows`) — `isMountpoint`
  (st_dev vs parent) and `fsType` (statfs f_type), moved out of
  storage.go unchanged. `int64(st.Type)` covers both Linux (int64) and
  darwin (uint32).
- `internal/nvr/storage_windows.go` — stubs that return an ERROR, not a
  zero value. A silent `false, nil` from isMountpoint would read as "not
  a mountpoint" and a silent `0, nil` from fsType as "not a network
  filesystem", converting a build failure into a runtime guard that
  quietly returns the wrong verdict. The nvr module is Linux-only by
  nature (Frigate/wyze-bridge Linux containers, host networking for TUTK,
  an NFS/SMB mount), so these are unreachable in practice.
- `internal/nvr/storage.go` — drops the `syscall` import; the probe
  constructors now reference the per-platform helpers. No behavior change
  on Linux/darwin.
- `.github/workflows/ci.yml` — the cross-compile matrix now also runs
  `go build ./...` per target, so a Unix-only syscall in any package
  fails CI instead of hiding behind the shipped binary's import graph.

## Testing

Verified for every release target before and after:
`CGO_ENABLED=0 GOOS={windows,linux,darwin} GOARCH={amd64,arm64} go build ./...`
— all pass; windows/amd64 and windows/arm64 previously failed.

`go test ./...` and `go vet ./...` green.

Closes #663
